refactored project to be more organized. Added a function in execute to be more flexible and shorter with new helper function
