+++ src/utils.c
@@ -13,80 +13,57 @@
 #include "minishell.h"
 #include "libft/libft.h"
 
-void execute_command(char *path, char **cmd, char **envp)
+char *get_env_value(char **envp, const char *name)
 {
-    pid_t pid;
-    int status;
-
-    pid = fork();
-    status = 0;
-    if (pid == 0)
+    int i;
+    size_t len;
+	
+	len = strlen(name);
+	i = 0;
+    while (envp[i])
     {
-        if (execve(path, cmd, envp) == -1)
+        if (ft_strncmp(envp[i], name, len) == 0 && envp[i][len] == '=')
+            return (envp[i] + len + 1);
+        i++;
+    }
+    return (NULL);
+}
+
+char **get_env_path(char **envp, const char *name)
+{
+    char *value;
+	
+	value = get_env_value(envp, name);
+    if (!value)
+        return (NULL);
+    return (ft_split(value, ':'));
+}
+
+char *get_path(char **envp, char **cmd)
+{
+    int i;
+    char **paths;
+    char *finalpath;
+    char *tmp;
+
+	i = 0;
+	paths = get_env_path(envp, "PATH");
+	finalpath = NULL;
+	tmp = NULL;
+    while (paths && paths[i])
+    {
+        tmp = ft_strcatrealloc(paths[i], "/");
+        if (tmp)
+            tmp = ft_strcatrealloc(tmp, cmd[0]);
+
+        if (tmp && access(tmp, F_OK) == 0)
         {
-            perror("execve");
-            _exit(1);
+            finalpath = ft_strdup(tmp);
+            free(tmp);
+            break;
         }
+        free(tmp);
+        i++;
     }
-    else if (pid > 0)
-    {
-        waitpid(pid, &status, 0);
-    }
-    else
-    {
-        perror("fork");
-    }
+    return (finalpath);
 }
-
-int custom_cd(char **envp, char **args)
-{
-    char *target_dir;
-
-	target_dir = NULL;
-    if (!args[1] || ft_strlen(args[1]) == 0)
-    {
-        target_dir = get_env_value(envp, "HOME");
-        if (!target_dir)
-        {
-            printf("cd: HOME not set\n");
-            return (1);
-        }
-    }
-    else
-    {
-        target_dir = args[1];
-    }
-    if (chdir(target_dir) != 0)
-	{
-		perror("cd");
-		return (1);
-	}
-    return (0);
-}
-
-int custom_exit(char **args)
-{
-    int status;
-    int num_len;
-
-    status = 0;
-    if (!args || !args[0])
-    {
-        printf("exit\n");
-        exit(status);
-    }
-    num_len = ft_is_str_num(args[0]);
-    if (num_len == 0 || num_len > 3) {
-        printf("exit: %s: numeric argument required\n", args[0]);
-        exit(255);
-    }
-    status = ft_atoi(args[0]);
-    if (status < 0 || status > 255) {
-        printf("exit\n");
-        exit(255);
-    }
-    printf("exit\n");
-    exit(status);
-    return 0;
-}
-
